requirements.txt
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
factory-boy==3.1.0
flake8
psycopg2-binary>=2.7,<3.0
celery==4.4.6
redis==3.5.3